Changed base weapons given to players
There are a few changes I'm looking to make over time to how retakes plays so that it brings money management and health into the mix when practicing retakes. In an actual game not every player would have full armor and health and the best weapon possible when the bomb has been planted. My goal is bring these variables into retakes.